refactor: moved the drag listeners to RQDraggableWidget abstract class
