+++ browser-extension/common/src/custom-elements/session-recording-widget/index.ts
@@ -1,6 +1,7 @@
 import styles from "./index.css";
 import { registerCustomElement, setInnerHTML } from "../utils";
 import BinIcon from "../../../resources/icons/bin.svg";
+import { RQDraggableWidget } from "../abstract-classes/draggable-widget";
 
 enum RQSessionRecordingWidgetEvent {
   STOP_RECORDING = "stop",
@@ -11,12 +12,12 @@ enum RQSessionRecordingWidgetEvent {
 const TAG_NAME = "rq-session-recording-widget";
 const DEFAULT_POSITION = { left: 30, bottom: 30 };
 
-class RQSessionRecordingWidget extends HTMLElement {
+class RQSessionRecordingWidget extends RQDraggableWidget {
   #shadowRoot;
-  #isDragging = false;
+  isDragging = false;
 
   constructor() {
-    super();
+    super(DEFAULT_POSITION);
     this.#shadowRoot = this.attachShadow({ mode: "closed" });
     setInnerHTML(this.#shadowRoot, this._getDefaultMarkup());
 
@@ -26,6 +27,7 @@ class RQSessionRecordingWidget extends HTMLElement {
 
   connectedCallback() {
     this.addListeners();
+    super.addDragListeners();
     this.show();
   }
 
@@ -47,71 +49,17 @@ class RQSessionRecordingWidget extends HTMLElement {
 
     this.addEventListener("hide", this.hide);
 
-    // allow widget to be draggable using mouse events
-    this.addEventListener("mousedown", (evt: MouseEvent) => {
-      evt.preventDefault();
-
-      let x = evt.clientX;
-      let y = evt.clientY;
-
-      const onMouseMove = (evt: MouseEvent) => {
-        evt.preventDefault();
-
-        this.#isDragging = true;
-
-        const dX = evt.clientX - x;
-        const dY = evt.clientY - y;
-
-        x = evt.clientX;
-        y = evt.clientY;
-
-        const xPos = Math.min(Math.max(this.offsetLeft + dX, 0), window.innerWidth - this.offsetWidth);
-        const yPos = Math.min(Math.max(this.offsetTop + dY, 0), window.innerHeight - this.offsetHeight);
-        this.moveToPostion({ top: yPos, left: xPos });
-      };
-
-      const onMouseUp = () => {
-        document.removeEventListener("mousemove", onMouseMove);
-        document.removeEventListener("mouseup", onMouseUp);
-      };
-
-      document.addEventListener("mousemove", onMouseMove);
-      document.addEventListener("mouseup", onMouseUp);
-    });
-
     this.#shadowRoot.addEventListener(
       "click",
       (evt) => {
-        if (this.#isDragging) {
+        if (this.isDragging) {
           // disable all clicks while widget is dragging
           evt.stopPropagation();
-          this.#isDragging = false;
+          this.isDragging = false;
         }
       },
       true
     );
-
-    window.addEventListener("resize", () => {
-      const boundingRect = this.getBoundingClientRect();
-
-      if (
-        boundingRect.left + boundingRect.width > window.innerWidth ||
-        boundingRect.top + boundingRect.height > window.innerHeight
-      ) {
-        this.moveToPostion(DEFAULT_POSITION);
-      }
-    });
-  }
-
-  moveToPostion(position: { top?: number; bottom?: number; left?: number; right?: number }) {
-    const getCSSPostionValue = (num?: number) => (typeof num !== "undefined" ? `${num}px` : "auto");
-
-    this.style.left = getCSSPostionValue(position.left);
-    this.style.top = getCSSPostionValue(position.top);
-    this.style.bottom = getCSSPostionValue(position.bottom);
-    this.style.right = getCSSPostionValue(position.right);
-
-    this.triggerEvent(RQSessionRecordingWidgetEvent.MOVED, position);
   }
 
   triggerEvent(name: RQSessionRecordingWidgetEvent, detail?: unknown) {
