+++ browser-extension/common/src/custom-elements/test-rule-widget/index.ts
@@ -2,6 +2,7 @@ import styles from "./index.css";
 import { registerCustomElement, setInnerHTML } from "../utils";
 import TickIcon from "../../../resources/icons/tick.svg";
 import CrossIcon from "../../../resources/icons/close.svg";
+import { RQDraggableWidget } from "../abstract-classes/draggable-widget";
 
 enum RQTestRuleWidgetEvent {
   VIEW_RESULTS = "view-results",
@@ -10,19 +11,21 @@ enum RQTestRuleWidgetEvent {
 const TAG_NAME = "rq-test-rule-widget";
 const DEFAULT_POSITION = { right: 10, top: 10 };
 
-class RQTestRuleWidget extends HTMLElement {
+class RQTestRuleWidget extends RQDraggableWidget {
+  isDragging: boolean = false;
   #shadowRoot;
-  #isDragging = false;
   #testRuleId: string;
 
   constructor() {
-    super();
+    super(DEFAULT_POSITION);
     this.#shadowRoot = this.attachShadow({ mode: "closed" });
     setInnerHTML(this.#shadowRoot, this._getDefaultMarkup());
   }
 
   connectedCallback() {
     this.setAttribute("draggable", "true");
+    super.addDragListeners();
+
     this.addListeners();
 
     this.#testRuleId = this.attributes.getNamedItem("rule-id")?.value;
@@ -55,68 +58,17 @@ class RQTestRuleWidget extends HTMLElement {
       }
     });
 
-    this.addEventListener("mousedown", (evt: MouseEvent) => {
-      evt.preventDefault();
-
-      let x = evt.clientX;
-      let y = evt.clientY;
-
-      const onMouseMove = (evt: MouseEvent) => {
-        evt.preventDefault();
-
-        this.#isDragging = true;
-
-        const dX = evt.clientX - x;
-        const dY = evt.clientY - y;
-
-        x = evt.clientX;
-        y = evt.clientY;
-
-        const xPos = Math.min(Math.max(this.offsetLeft + dX, 0), window.innerWidth - this.offsetWidth);
-        const yPos = Math.min(Math.max(this.offsetTop + dY, 0), window.innerHeight - this.offsetHeight);
-        this.moveToPostion({ top: yPos, left: xPos });
-      };
-
-      const onMouseUp = () => {
-        document.removeEventListener("mousemove", onMouseMove);
-        document.removeEventListener("mouseup", onMouseUp);
-      };
-
-      document.addEventListener("mousemove", onMouseMove);
-      document.addEventListener("mouseup", onMouseUp);
-    });
-
     this.#shadowRoot.addEventListener(
       "click",
       (evt) => {
-        if (this.#isDragging) {
+        if (this.isDragging) {
           // disable all clicks while widget is dragging
           evt.stopPropagation();
-          this.#isDragging = false;
+          this.isDragging = false;
         }
       },
       true
     );
-
-    window.addEventListener("resize", () => {
-      const boundingRect = this.getBoundingClientRect();
-
-      if (
-        boundingRect.left + boundingRect.width > window.innerWidth ||
-        boundingRect.top + boundingRect.height > window.innerHeight
-      ) {
-        this.moveToPostion(DEFAULT_POSITION);
-      }
-    });
-  }
-
-  moveToPostion(position: { top?: number; bottom?: number; left?: number; right?: number }) {
-    const getCSSPostionValue = (num?: number) => (typeof num !== "undefined" ? `${num}px` : "auto");
-
-    this.style.left = getCSSPostionValue(position.left);
-    this.style.top = getCSSPostionValue(position.top);
-    this.style.bottom = getCSSPostionValue(position.bottom);
-    this.style.right = getCSSPostionValue(position.right);
   }
 
   triggerEvent(name: RQTestRuleWidgetEvent, detail?: unknown) {
